requirements.txt
numpy
tk
Pillow
torch>=1.12.1